Fixed compiler warnings on 64bit GCC-5.2.1

+++ src/nvthread/ThreadPool.cpp
@@ -81,7 +81,7 @@ ThreadPool::ThreadPool()
 
     nvCompilerWriteBarrier(); // @@ Use a memory fence?
 
-    for (uint i = 0; i < workerCount; i++) {
+    for (uintptr_t i = 0; i < workerCount; i++) {
         workers[i].start(workerFunc, (void *)i);
     }
